NewsCard refactor in progress
