+++ src/components/App/App.jsx
@@ -83,6 +83,7 @@ function App() {
     setIsLoggedIn(false);
     localStorage.setItem("token", "");
     setToken("");
+    setNewsCards([]);
   }
 
   useEffect(() => {
