+++ src/components/NewsCardReadOnly/NewsCardReadOnly.jsx
@@ -1,0 +1,16 @@
+import bookmarkUnmarked from "../../assets/bookmark-unmarked.png";
+import bookmarkMarked from "../../assets/bookmark-marked.png";
+import bookmarkContainer from "../../assets/bookmark-container.png";
+
+// function NewsCardReadOnly () {
+//     return (
+//            <div className="newsCard__bookmark-container">
+//               <img
+//                 className="newsCard__bookmark newsCard__bookmark_unmarked"
+//                 src={bookmarkUnmarked}
+//                 alt="bookmark unmarked"
+//               />
+//                <div className="newsCard__login-message">Sign in to save</div>
+//               </div>
+//     )
+// }
