+++ src/components/Results/Results.jsx
@@ -26,6 +26,8 @@ function Results({ newsCards, isLoading, noResults, isLoggedIn }) {
     );
   }
 
+  console.log("newsCards in Results:", newsCards);
+
   return <div className="results">{content}</div>;
 
   // return (
